+++ CHANGELOG.md
@@ -7,9 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- All Sales Admin users are notified via email when a new quote is requested
+- All users who have interacted with a quote are notified via email when the quote is updated
+- Default `QuoteCreated` and `QuoteUpdated` email templates to support the above
+
 ## [1.1.0] - 2022-01-28
 
 ### Added
+
 - Added on getQuote/getQuote a checker if the property b2b-quote-graphQL exists in the orderForm if not then it creates a new property there.
 
 ## [1.0.3] - 2022-01-06